Temp: shake to pay 설정
