+++ pages/_app.tsx
@@ -6,24 +6,8 @@ import { Toaster } from 'react-hot-toast';
 import '../styles/base/_variable.scss';
 import '../styles/index.scss';
 import { useEffect } from 'react';
-function shakeEventDidOccur() {
-  //put your own code here etc.
-  alert('shake!');
-}
+
 const MyApp = ({ Component, pageProps }: AppProps) => {
-  var Shake = require('shake.js');
-  useEffect(() => {
-    var myShakeEvent = new Shake({
-      threshold: 15, // optional shake strength threshold
-      timeout: 1000, // optional, determines the frequency of event generation
-    });
-
-    myShakeEvent.start();
-    window.addEventListener('shake', shakeEventDidOccur, false);
-
-    //function to call when shake occurs
-  }, []);
-
   return (
     <RecoilRoot>
       <ThemeProvider>
