+++ src/store/atom/optionState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const shakeToPayState = atom({
+  key: 'shakeToPayState',
+  default: false,
+});
